Remove smem utils and factor out reduce-scan common
